Complete reimplementation using MenuComponent
Make PancakeHouseMenu and the other original menus into classes with single
static methods that return a MenuComponent. This allows the code that
seeds the menus to remain in those classes. The demo creates the components,
tosses them into a unified menu, and gives them to a new waitress like
the example.